+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-16 16:15:44
-// Commit: style(perf): harmonize light and dark theme palette for perf
+// Updated at: 2026-04-16 16:42:06
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '428';
-  final String lastMessage = 'style(perf): harmonize light and dark theme palette for perf';
-  final DateTime timestamp = DateTime.parse('2026-04-16T16:15:44+05:00');
+  final String commitRef = '430';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-04-16T16:42:06+05:00');
 }